requirements.txt
mcp>=1.0.0
Pillow>=9.0.0